refactor terraform wrapper script into its own file
